Modify numberOfYears from int to float

+++ src/Context/Backend/Calculator/CalculateStrategy/FinalAmountWithRegularPayments.php
@@ -33,7 +33,7 @@ class FinalAmountWithRegularPayments implements CalculateStrategyInterface
             (float)$input->getInitialAmount(),
             (float)$input->getRegularPayment(),
             $input->getNumberOfRegularPaymentsPerYear(),
-            (int)$input->getNumberOfYears(),
+            (float)$input->getNumberOfYears(),
             (float)$input->getInterestRatePerYear(),
             round($finalAmount, 2)
         );
